Simple-deployment v. 5.0.0 - Allow OK domain in Cloud (#133)
Simple-deployment v. 5.0.0

New logic for determining what ingress and cluster issuer should be used.
The previous logic determined what cluster you where deploying in based on the hostname you chose for your service.
This has been phased out, to support only using the ok.dk domain for our services in the future, regardless of where the service is deployed.

Without breaking support for overwriting ingressClassName or clusterIssuer in the services' values file.
Clusterissuer can be written manually or by using the clusterIssuer variable.
With private by default on Gen 2 kubernetes platform and support for public to comply with CISO.
The ingressClassName and clusterIssuer is adapting to the cluster your service is deployed in.
This PR includes unittests for the different cases that we now support in Simple-deployment.

The different scenarios in this release have been tested in templates.

The ReadMe has been updated to explain support for private ingress. Also showing how to make your service public on Gen 2.

Breaking changes
Important! The cluster for deployment is no longer be derived from the domain name.
Developers are required to specify what cluster they want to deploy their services in. This is done using the cluster variable.
The cluster variable is case insensitive.

This will choose the correct default ingressClassName and clusterIssuer for our cloud environment.
ingress:
enable: true
host: example.test.ok.dk
cluster: Cloud

This will chose the correct default ingressClassName and clusterIssuer for our On-premises environment.
ingress:
enable: true
host: example.test.ok.dk
cluster: Gen2